Code cleanup in reflect.go
reflect.go's field function was pretty large so I tried to break it
into a few more logical smaller functions to reduce complexity. I'm
still only testing the field function and relying on it calling
other functions to test them, as the only thing I really care about
is that the fields function works as expected however that gets
done.